Compiled some files

+++ Userland/UserModule/standard_library/mystdlib.c
@@ -50,7 +50,7 @@ static int isDigit(int digit){
 }
 
 // Transforms an integer to string and stores it on target. Returns final string length
-size_t numToStr(int value, char* target, size_t base){
+size_t numToStr(long value, char* target, size_t base){
 	int digit;
 	int sign = 1;		// 0 if negative, 1 if positive
 	int i = -1, j = 0;
